refactor: PieceFactory를 PieceType으로 대신함
